RF | move testCreatingProjectWithInvalidName
